+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Scheduled maintenance fees zijn zichtbaar in de app 1 week voordat ze effectief betaald moeten worden
 - Mazebank Arena interior toegevoegd met mogelijkheid om te veranderen door staff
 - Emote toegevoegd
+- Command `reportHouserobbery` toegevoegd om je huidige houserobbery locatie te melden as scuffed (door water etc)
 
 ### Changed
 - Porto crafting requirement aangepast